Persistent Remap Key Modifier

+++ Auto/Auto/Constants.cs
@@ -6,12 +6,12 @@ public class Constants
 {
     public const int WH_KEYBOARD_LL = 13;
     public const int WH_MOUSE_LL = 14;
-    public static readonly IntPtr WM_KEYDOWN = (IntPtr)0x0100;
-    public static readonly IntPtr WM_KEYUP = (IntPtr)0x0101;
-    public static readonly IntPtr WM_SYSKEYDOWN = (IntPtr)0x0104;
-    public static readonly IntPtr WM_SYSKEYUP = (IntPtr)0x0105;
-    public static readonly IntPtr KEY_DOWN_UP = (IntPtr)0x0000;
-    public static readonly IntPtr WM_LBUTTONDOWN = (IntPtr)0x0201;
+    public static readonly nint WM_KEYDOWN = 0x0100;
+    public static readonly nint WM_KEYUP = 0x0101;
+    public static readonly nint WM_SYSKEYDOWN = 0x0104;
+    public static readonly nint WM_SYSKEYUP = 0x0105;
+    public static readonly nint KEY_DOWN_UP = 0x0000;
+    public static readonly nint WM_LBUTTONDOWN = 0x0201;
     public static readonly uint MOUSEEVENTF_LEFTDOWN = 0x0002;
     public static readonly uint MOUSEEVENTF_LEFTUP = 0x0004;
     public static readonly int IGNORE_INPUT = 0xffffff;
@@ -23,7 +23,7 @@ public class Constants
         public ushort wScan;
         public uint dwFlags;
         public uint time;
-        public IntPtr dwExtraInfo;
+        public nint dwExtraInfo;
     }
 
     [StructLayout(LayoutKind.Sequential)]
@@ -34,7 +34,7 @@ public class Constants
         public uint mouseData;
         public uint dwFlags;
         public uint time;
-        public IntPtr dwExtraInfo;
+        public nint dwExtraInfo;
     }
 
     [StructLayout(LayoutKind.Sequential)]
